+++ src/common/factories/mongoose-model.factory.ts
@@ -1,0 +1,13 @@
+import { Schema } from 'mongoose';
+import { ModelDefinition } from '@nestjs/mongoose';
+
+export class MongooseModelFactory {
+
+    static create(name: string, schema: Schema): ModelDefinition {
+        return {
+            name,
+            schema,
+            collection: name
+        };
+    }
+}