+++ src/modules/dish/dish.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { DishService } from './dish.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { DishController } from './dish.controller';
-import { DISH_COMMENT_MODEL, DISH_MODEL, DISH_RATING_MODEL, SEARCH_QUERY_MODEL } from '../../constants/models.constant';
+import { dishCommentModel, dishModel, dishRatingModel, userSearchQueryModel } from '../../common/definitions/mongoose-model.definitions';
 import { RedisModule } from '../redis/redis.module';
 import { DishRepository } from '../../mongodb/repositories/dish.repository';
 import { JwtManagerModule } from '../jwt-manager/jwt-manager.module';
 import { JwtManagerService } from '../jwt-manager/jwt-manager.service';
 import { IngredientModule } from '../ingredient/ingredient.module';
-import { SearchQueryRepository } from '../../mongodb/repositories/search-query.repository';
+import { UserSearchQueryRepository } from '../../mongodb/repositories/user-search-query.repository';
 import { TranslationModule } from '../translation/translation.module';
 import { DishCommentRepository } from '../../mongodb/repositories/dish-comment.repository';
 import { DishRatingRepository } from '../../mongodb/repositories/dish-rating.repository';
@@ -16,7 +16,7 @@ import { ExternalApiModule } from '../api/external-api.module';
 
 @Module({
     imports: [
-        MongooseModule.forFeature([DISH_MODEL, DISH_COMMENT_MODEL, DISH_RATING_MODEL, SEARCH_QUERY_MODEL]),
+        MongooseModule.forFeature([dishModel, dishCommentModel, dishRatingModel, userSearchQueryModel]),
         RedisModule,
         JwtManagerModule,
         IngredientModule,
@@ -24,6 +24,6 @@ import { ExternalApiModule } from '../api/external-api.module';
         TranslationModule
     ],
     controllers: [DishController],
-    providers: [DishService, DishRepository, DishCommentRepository, DishRatingRepository, SearchQueryRepository, JwtManagerService],
+    providers: [DishService, DishRepository, DishCommentRepository, DishRatingRepository, UserSearchQueryRepository, JwtManagerService],
 })
 export class DishModule {}