+++ src/mongodb/repositories/dish-comment.repository.ts
@@ -1,13 +1,13 @@
 import { AbstractRepository } from './abstract.repository';
 import { DishCommentDocument } from '../documents/dish-comment.document';
 import { InjectModel } from '@nestjs/mongoose';
-import { models } from '../../constants/models.constant';
+import { dishCommentModel } from '../../common/definitions/mongoose-model.definitions';
 import { Model } from 'mongoose';
 import { CreateDishCommentDto } from '../../modules/dish/dish.dto';
 
 export class DishCommentRepository extends AbstractRepository<DishCommentDocument, CreateDishCommentDto> {
 
-    constructor(@InjectModel(models.DISH_COMMENT_MODEL) model: Model<DishCommentDocument>) {
+    constructor(@InjectModel(dishCommentModel.name) model: Model<DishCommentDocument>) {
         super(model);
     }