+++ src/mongodb/repositories/user.action.repository.ts
@@ -2,11 +2,11 @@ import { AbstractRepository } from './abstract.repository';
 import { UserActionDocument } from '../documents/user-action.document';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { models } from '../../constants/models.constant';
+import { userActionModel } from '../../common/definitions/mongoose-model.definitions';
 
 export class UserActionRepository extends AbstractRepository<UserActionDocument, { userId: string, type: string }> {
 
-    constructor(@InjectModel(models.USER_ACTION_MODEL) model: Model<UserActionDocument>) {
+    constructor(@InjectModel(userActionModel.name) model: Model<UserActionDocument>) {
         super(model);
     }
 }